adding authentication and autherization with error handling

+++ controllers/admin.js
@@ -1,3 +1,4 @@
+const { validationResult } = require("express-validator");
 const Product = require("../models/product");
 
 exports.getAddProduct = (req, res, next) => {
@@ -5,12 +6,31 @@ exports.getAddProduct = (req, res, next) => {
     docTitle: "Add Product",
     path: "/admin/add-product",
     editing: false,
-    isAuthenticated: req.session.isLoggedIn,
+    hasError: false,
+    errorMessage: null,
+    validationErrors: [],
   });
 };
 
 exports.postAddProduct = (req, res, next) => {
   const { title, price, description, imageUrl } = req.body;
+  const errors = validationResult(req);
+  if (!errors.isEmpty()) {
+    return res.status(422).render("admin/edit-product", {
+      docTitle: "Add Product",
+      path: "/admin/add-product",
+      editing: false,
+      hasError: true,
+      errorMessage: errors.array()[0].msg,
+      validationErrors: errors.array(),
+      product: {
+        title: title,
+        price: price,
+        description: description,
+        imageUrl: imageUrl,
+      },
+    });
+  }
   const product = new Product({
     title,
     price,
@@ -18,7 +38,14 @@ exports.postAddProduct = (req, res, next) => {
     imageUrl,
     userId: req.user._id,
   });
-  product.save().then(() => res.redirect("/admin/products"));
+  product
+    .save()
+    .then(() => res.redirect("/admin/products"))
+    .catch((err) => {
+      const error = new Error(err);
+      error.httpStatusCode = 500;
+      next(error);
+    });
 };
 
 exports.getEditProduct = (req, res, next) => {
@@ -35,6 +62,9 @@ exports.getEditProduct = (req, res, next) => {
       docTitle: "Edit Product",
       path: "/admin/edit-product",
       editing: editMode,
+      hasError: false,
+      errorMessage: null,
+      validationErrors: [],
       product,
     });
   });
@@ -42,33 +72,65 @@ exports.getEditProduct = (req, res, next) => {
 
 exports.postEditProduct = (req, res, next) => {
   const { productId, title, price, description, imageUrl } = req.body;
-  Product.updateOne({ _id: productId }, { title, price, description, imageUrl })
+  const errors = validationResult(req);
+  if (!errors.isEmpty()) {
+    return res.status(422).render("admin/edit-product", {
+      docTitle: "Edit Product",
+      path: "/admin/add-product",
+      editing: true,
+      hasError: true,
+      errorMessage: errors.array()[0].msg,
+      validationErrors: errors.array(),
+      product: {
+        title: title,
+        price: price,
+        description: description,
+        imageUrl: imageUrl,
+        _id: productId,
+      },
+    });
+  }
+  Product.updateOne(
+    { _id: productId, userId: req.user._id },
+    { title, price, description, imageUrl }
+  )
     .then(() => {
       console.log("PRODUCT UPDATED!");
       res.redirect("/admin/products");
     })
-    .catch((err) => console.log(err));
+    .catch((err) => {
+      const error = new Error(err);
+      error.httpStatusCode = 500;
+      next(error);
+    });
 };
 
 exports.getProducts = (req, res, next) => {
-  Product.find()
+  Product.find({ userId: req.user._id })
     .then((products) => {
       res.render("admin/products", {
         docTitle: "All Products",
         products,
         path: "/admin/products",
-        isAuthenticated: req.session.isLoggedIn,
       });
     })
-    .catch((err) => console.log(err));
+    .catch((err) => {
+      const error = new Error(err);
+      error.httpStatusCode = 500;
+      next(error);
+    });
 };
 
 exports.postDeleteProduct = (req, res, next) => {
   const { productId } = req.body;
-  Product.deleteOne({ _id: productId })
+  Product.deleteOne({ _id: productId, userId: req.user._id })
     .then(() => {
       console.log("PRODUCT DELETED!");
       res.redirect("/admin/products");
     })
-    .catch((err) => console.log(err));
+    .catch((err) => {
+      const error = new Error(err);
+      error.httpStatusCode = 500;
+      next(error);
+    });
 };
